Initial work to use config with track extractor.
