+++ track/trackextractor.py
@@ -1,15 +1,24 @@
 """
-Author Matthew Aitchison
+classifier-pipeline - this is a server side component that manipulates cptv
+files and to create a classification model of animals present
+Copyright (C) 2018, The Cacophony Project
 
-Date December 2017
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
 
-Analysis a CPTV video file and extracts tracks for any moving objects.  These tracks are then saved out to a
-HDF5 file for use in training a neural net.
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <http://www.gnu.org/licenses/>.
 """
 
 import pytz
 import datetime
-from collections import namedtuple
 import os
 import pytz
 
@@ -20,355 +29,18 @@ import scipy.ndimage
 import logging
 
 from ml_tools.tools import get_image_subsection, blosc_zstd
-from ml_tools.tools import Rectangle
 from ml_tools.trackdatabase import TrackDatabase
 from ml_tools import tools
 from ml_tools.dataset import TrackChannels
 
+from track.track import Track
+from track.region import Region
+from track.framebuffer import FrameBuffer
+
 from cptv import CPTVReader
-
-__import__('tables')  # <-- import PyTables; __import__ so that linters don't complain
-
-
-class Region(Rectangle):
-    """ Region is a rectangle extended to support mass. """
-    def __init__(self, topleft_x, topleft_y, width, height, mass=0, pixel_variance=0, id=0, frame_index=0, was_cropped=False):
-        super().__init__(topleft_x, topleft_y, width, height)
-        # number of active pixels in region
-        self.mass = mass
-        # how much pixels in this region have changed since last frame
-        self.pixel_variance = pixel_variance
-        # an identifier for this region
-        self.id = id
-        # frame index from clip
-        self.frame_index = frame_index
-        # if this region was cropped or not
-        self.was_cropped = was_cropped
-
-    def copy(self):
-        return Region(
-            self.x, self.y, self.width, self.height, self.mass, self.pixel_variance, self.id, self.frame_index,
-            self.was_cropped
-        )
-
-TrackMovementStatistics = namedtuple(
-    'TrackMovementStatistics',
-    'movement max_offset score average_mass median_mass duration delta_std'
-)
-TrackMovementStatistics.__new__.__defaults__ = (0,) * len(TrackMovementStatistics._fields)
-
-
-class Track:
-    """ Bounds of a tracked object over time. """
-
-    # keeps track of which id number we are up to.
-    _track_id = 1
-
-    def __init__(self, id=None):
-        """
-        Creates a new Track.
-        :param id: id number for track, if not specified is provided by an auto-incrementer
-        """
-
-        # used to uniquely identify the track
-        if not id:
-            self.id = self._track_id
-            self._track_id += 1
-        else:
-            self.id = id
-
-        # frame number this track starts at
-        self.start_frame = 0
-        # datetime this track starts
-        self.start_time = None
-        # our bounds over time
-        self.bounds_history = []
-        # number frames since we lost target.
-        self.frames_since_target_seen = 0
-        # our current estimated horizontal velocity
-        self.vel_x = 0
-        # our current estimated vertical velocity
-        self.vel_y = 0
-
-        # the tag for this track
-        self.tag = "unknown"
-
-    def add_frame(self, bounds: Region):
-        """
-        Adds a new point in time bounds and mass to track
-        :param bounds: new bounds region
-        """
-        self.bounds_history.append(bounds.copy())
-        self.frames_since_target_seen = 0
-
-        if len(self) >= 2:
-            self.vel_x = self.bounds_history[-1].mid_x - self.bounds_history[-2].mid_x
-            self.vel_y = self.bounds_history[-1].mid_y - self.bounds_history[-2].mid_y
-        else:
-            self.vel_x = self.vel_y = 0
-
-    def add_blank_frame(self):
-        """ Maintains same bounds as previously, does not reset framce_since_target_seen counter """
-        self.bounds_history.append(self.bounds.copy())
-        self.bounds.mass = 0
-        self.bounds.pixel_variance = 0
-        self.bounds.frame_index += 1
-        self.vel_x = self.vel_y = 0
-
-    def get_stats(self):
-        """
-        Returns statistics for this track, including how much it moves, and a score indicating how likely it is
-        that this is a good track.
-        :return: a TrackMovementStatistics record
-        """
-
-        if len(self) <= 1:
-            return TrackMovementStatistics()
-
-        # get movement vectors
-        mass_history = [int(bound.mass) for bound in self.bounds_history]
-        variance_history = [bound.pixel_variance for bound in self.bounds_history]
-        mid_x = [bound.mid_x for bound in self.bounds_history]
-        mid_y = [bound.mid_y for bound in self.bounds_history]
-        delta_x = [mid_x[0] - x for x in mid_x]
-        delta_y = [mid_y[0] - y for y in mid_y]
-        vel_x = [cur - prev for cur, prev in zip(mid_x[1:], mid_x[:-1])]
-        vel_y = [cur - prev for cur, prev in zip(mid_y[1:], mid_y[:-1])]
-
-        movement = sum((vx ** 2 + vy ** 2) ** 0.5 for vx, vy in zip(vel_x, vel_y))
-        max_offset = max((dx ** 2 + dy ** 2) ** 0.5 for dx, dy in zip(delta_x, delta_y))
-
-        # the standard deviation is calculated by averaging the per frame variances.
-        # this ends up being slightly different as I'm using /n rather than /(n-1) but that
-        # shouldn't make a big difference as n = width*height*frames which is large.
-        delta_std = float(np.mean(variance_history)) ** 0.5
-
-        movement_points = (movement ** 0.5) + max_offset
-        delta_points = delta_std * 25.0
-        score = min(movement_points,100) + min(delta_points, 100)
-
-        stats = TrackMovementStatistics(
-            movement=float(movement),
-            max_offset=float(max_offset),
-            average_mass=float(np.mean(mass_history)),
-            median_mass=float(np.median(mass_history)),
-            duration=len(self) / 9.0,
-            delta_std=float(delta_std),
-            score=float(score)
-        )
-
-        return stats
-
-    def smooth(self, frame_bounds: Rectangle):
-        """
-        Smooths out any quick changes in track dimensions
-        :param frame_bounds The boundaries of the video frame.
-        """
-        if len(self.bounds_history) == 0:
-            return
-
-        new_bounds_history = []
-        prev_frame = self.bounds_history[0]
-        current_frame = self.bounds_history[0]
-        next_frame = self.bounds_history[1]
-
-        for i in range(len(self.bounds_history)):
-
-            prev_frame = self.bounds_history[max(0, i-1)]
-            current_frame = self.bounds_history[i]
-            next_frame = self.bounds_history[min(len(self.bounds_history)-1, i+1)]
-
-            frame_x = current_frame.mid_x
-            frame_y = current_frame.mid_y
-            frame_width = (prev_frame.width + current_frame.width + next_frame.width) / 3
-            frame_height = (prev_frame.height + current_frame.height + next_frame.height) / 3
-            frame = Region(int(frame_x - frame_width / 2), int(frame_y - frame_height / 2), int(frame_width), int(frame_height))
-            frame.crop(frame_bounds)
-
-            new_bounds_history.append(frame)
-
-        self.bounds_history = new_bounds_history
-
-    def trim(self):
-        """
-        Removes empty frames from start and end of track
-        """
-        mass_history = [int(bound.mass) for bound in self.bounds_history]
-
-        start = 0
-        while start < len(self) and mass_history[start] <= 2:
-            start += 1
-        end = len(self)-1
-        while end > 0 and mass_history[end] <= 2:
-            end -= 1
-
-        if end < start:
-            end = start
-
-        self.start_time += datetime.timedelta(seconds=start / 9.0)
-        self.start_frame += start
-        self.bounds_history = self.bounds_history[start:end-1]
-
-    def get_track_region_score(self, region: Region):
-        """
-        Calculates a score between this track and a region of interest.  Regions that are close the the expected
-        location for this track are given high scores, as are regions of a similar size.
-        """
-        expected_x = int(self.bounds.mid_x + self.vel_x)
-        expected_y = int(self.bounds.mid_y + self.vel_y)
-
-        distance = ((region.mid_x - expected_x) ** 2 + (region.mid_y - expected_y) ** 2) ** 0.5
-
-        # ratio of 1.0 = 20 points, ratio of 2.0 = 10 points, ratio of 3.0 = 0 points.
-        # area is padded with 50 pixels so small regions don't change too much
-        size_difference = (abs(region.area - self.bounds.area) / (self.bounds.area+50)) * 100
-
-        return distance, size_difference
-
-    def get_overlap_ratio(self, other_track, threshold = 0.05):
-        """
-        Checks what ratio of the time these two tracks overlap.
-        :param other_track: the other track to compare with
-        :param threshold: how much frames must be overlapping to be counted
-        :return: the ratio of frames that overlap
-        """
-
-        if len(self) == 0 or len(other_track) == 0:
-            return 0.0
-
-        start = max(self.start_frame, other_track.start_frame)
-        end = min(self.start_frame + len(self), other_track.start_frame + len(other_track))
-
-        frames_overlapped = 0
-
-        for pos in range(start, end+1):
-            our_index = pos - self.start_frame
-            other_index = pos - other_track.start_frame
-            if our_index >= 0 and other_index >= 0 and our_index < len(self) and other_index < len(other_track):
-                our_bounds = self.bounds_history[our_index]
-                other_bounds = other_track.bounds_history[other_index]
-                overlap = our_bounds.overlap_area(other_bounds) / our_bounds.area
-                if overlap >= threshold:
-                    frames_overlapped += 1
-
-        return frames_overlapped / len(self)
-
-    @property
-    def mass(self):
-        return self.bounds_history[-1].mass
-
-    @property
-    def bounds(self) -> Region:
-        return self.bounds_history[-1]
-
-    @property
-    def duration(self):
-        """ Duration of track in seconds. """
-        return len(self) / 9.0
-
-    @property
-    def end_time(self):
-        return self.start_time + datetime.timedelta(seconds=self.duration)
-
-    def __repr__(self):
-        return "Track:{} frames".format(len(self))
-
-    def __len__(self):
-        return len(self.bounds_history)
-
-
-class BackgroundAnalysis:
-    """ Stores background analysis statistics. """
-    def __init__(self):
-        self.threshold = None
-        self.average_delta = None
-        self.max_temp = None
-        self.min_temp = None
-        self.mean_temp = None
-        self.background_deviation = None
-
-class FrameBuffer:
-    """ Stores entire clip in memory, required for some operations such as track exporting. """
-    def __init__(self):
-        self.thermal = None
-        self.filtered = None
-        self.delta = None
-        self.mask = None
-        self.flow = None
-        self.reset()
-
-    @property
-    def has_flow(self):
-        return self.flow is not None and len(self.flow) != 0
-
-    def generate_flow(self, opt_flow, flow_threshold=40):
-        """
-        Generate optical flow from thermal frames
-        :param opt_flow: An optical flow algorithm
-        """
-
-        self.flow = []
-
-        height, width = self.filtered[0].shape
-        flow = np.zeros([height, width, 2], dtype=np.float32)
-
-        current = None
-        for frame in self.thermal:
-            frame = np.float32(frame)
-            # strong filtering helps with the optical flow.
-            threshold = np.median(frame) + flow_threshold
-            next = np.uint8(np.clip(frame - threshold, 0, 255))
-
-            if current is not None:
-                # for some reason openCV spins up lots of threads for this which really slows things down, so we
-                # cap the threads to 2
-                cv2.setNumThreads(2)
-                flow = opt_flow.calc(current, next, flow)
-
-            current = next
-
-            # scale up the motion vectors so that we get some additional precision
-            # but also make sure they fit within an int16
-            scaled_flow = np.clip(flow * 256, -16000, 16000)
-            self.flow.append(scaled_flow)
-
-    def reset(self):
-        """
-        Empties buffer
-        """
-        self.thermal = []
-        self.filtered = []
-        self.delta = []
-        self.mask = []
-        self.flow = []
-
-    def __len__(self):
-        return len(self.thermal)
 
 class TrackExtractor:
     """ Extracts tracks from a stream of frames. """
-
-    # enables fast BLOSC compression (requires plugin)
-    ENABLE_COMPRESSION = False
-
-    # includes the filtered channel when exporting.  This is typically not used.  If compression is enabled the
-    # filesize can be reduced by not including it.
-    INCLUDE_FILTERED_CHANNEL = True
-
-    # auto threshold needs to find a near maximum value to calculate the threshold level
-    # a better solution might be the mean of the max of each frame?
-    THRESHOLD_PERCENTILE = 99.9
-
-    # measures how different pixels are from estimated background, if they are on average less different than this then
-    # video is considered to be static.
-    STATIC_BACKGROUND_THRESHOLD = 4.0
-
-    # any clips with a mean temperature hotter than this will be excluded
-    MAX_MEAN_TEMPERATURE_THRESHOLD = 10000
-
-    # any clips with a temperature dynamic range greater than this will be excluded
-    MAX_TEMPERATURE_RANGE_THRESHOLD = 10000
-
     # number of pixels around object to pad.
     FRAME_PADDING = 6
 
@@ -386,7 +58,11 @@ class TrackExtractor:
     # when enabled smooths tracks so that track dimensions do not change too quickly.
     TRACK_SMOOTHING = False
 
-    def __init__(self):
+    def __init__(self, trackconfig):
+
+        self.config = trackconfig
+
+        self.compression = blosc_zstd if self.config.enable_compression else None
 
         # start time of video
         self.video_start_time = None
@@ -479,7 +155,7 @@ class TrackExtractor:
 
         # first we get the background.  This requires reading the entire source into memory.
         background, background_stats = self.analyse_background(frames)
-        is_static_background = background_stats.background_deviation < self.STATIC_BACKGROUND_THRESHOLD
+        is_static_background = background_stats.background_deviation < self.config.static_background_threshold
 
         self.stats['threshold'] = background_stats.threshold
         self.stats['average_background_delta'] = background_stats.background_deviation
@@ -504,12 +180,12 @@ class TrackExtractor:
             return False
 
         # don't process clips that are too hot.
-        if self.MAX_MEAN_TEMPERATURE_THRESHOLD and background_stats.mean_temp > self.MAX_MEAN_TEMPERATURE_THRESHOLD:
+        if self.config.max_mean_temperature_threshold and background_stats.mean_temp > self.config.max_mean_temperature_threshold:
             self.reject_reason = "Mean temp too high {}".format(background_stats.mean_temp)
             return False
 
         # don't process clips with too large of a temperature difference
-        if self.MAX_TEMPERATURE_RANGE_THRESHOLD and (background_stats.max_temp - background_stats.min_temp > self.MAX_TEMPERATURE_RANGE_THRESHOLD):
+        if self.config.max_temperature_range_threshold and (background_stats.max_temp - background_stats.min_temp > self.config.max_temperature_range_threshold):
             self.reject_reason = "Temp delta too high {}".format(background_stats.max_temp - background_stats.min_temp)
             return False
 
@@ -538,7 +214,7 @@ class TrackExtractor:
         self.filter_tracks()
 
         # apply smoothing if required
-        if self.TRACK_SMOOTHING and len(frames) > 0:
+        if self.config.track_smoothing and len(frames) > 0:
             frame_height, frame_width = frames[0].shape
             for track in self.tracks:
                 track.smooth(Rectangle(0,0,frame_width, frame_height))
@@ -622,11 +298,11 @@ class TrackExtractor:
                 channels = self.get_track_channels(track, i)
 
                 # zero out the filtered channel
-                if not self.INCLUDE_FILTERED_CHANNEL:
+                if not self.config.include_filtered_channel:
                     channels[TrackChannels.filtered] = 0
                 track_data.append(channels)
             track_id = track_number+1
-            database.add_track(clip_id, track_id, track_data, track, opts=blosc_zstd if self.ENABLE_COMPRESSION else None)
+            database.add_track(clip_id, track_id, track_data, track, opts=self.compression)
 
     def get_track_channels(self, track: Track, frame_number):
         """
@@ -724,8 +400,8 @@ class TrackExtractor:
             track.frames_since_target_seen += 1
             track.add_blank_frame()
 
-        # remove any tracks that have not seen their target in 9 frames
-        self.active_tracks = [track for track in self.active_tracks if track.frames_since_target_seen < self.DELETE_LOST_TRACK_FRAMES]
+        # remove any tracks that have not seen their target in a while
+        self.active_tracks = [track for track in self.active_tracks if track.frames_since_target_seen < self.config.remove_track_after_frames]
 
     def filter_tracks(self):
 
@@ -820,7 +496,7 @@ class TrackExtractor:
         labels, mask, stats, centroids = cv2.connectedComponentsWithStats(thresh)
 
         # we enlarge the rects a bit, partly because we eroded them previously, and partly because we want some context.
-        padding = self.FRAME_PADDING
+        padding = self.config.frame_padding
 
         # find regions of interest
         regions = []
@@ -835,13 +511,13 @@ class TrackExtractor:
             )
 
             old_region = region.copy()
-            if self.CROPPED_REGIONS_STRATEGY == "all":
+            if self.config.cropped_regions_strategy == "all":
                 # in this case we just clip the region to the edges of the screen
                 region.left = max(region.left, 0)
                 region.top = max(region.top, 0)
                 region.right = min(region.right, frame_width - 1)
                 region.bottom = min(region.bottom, frame_height - 1)
-            elif self.CROPPED_REGIONS_STRATEGY == "cautious":
+            elif self.config.cropped_regions_strategy == "cautious":
                 # keep cropped regions if they have a reasonable size
                 region.left = max(region.left, 0)
                 region.top = max(region.top, 0)
@@ -854,14 +530,14 @@ class TrackExtractor:
                 if crop_width_fraction > 0.25 or crop_height_fraction > 0.25:
                     continue
 
-            elif self.CROPPED_REGIONS_STRATEGY == "none":
+            elif self.config.cropped_regions_strategy == "none":
                 # all regions that touch the side of the screen are removed
                 if (region.left < 0) or (region.right > frame_width) or (region.top < 0) or (region.top < 0) or (region.bottom > frame_height):
                     continue
             else:
                 raise ValueError(
                     "Invalid mode for CROPPED_REGIONS_STRATEGY, expected ['all','cautious','none'] but found {}".format(
-                        self.CROPPED_REGIONS_STRATEGY))
+                        self.config.cropped_regions_strategy))
             region.was_cropped = str(old_region) != str(region)
 
             if delta_frame is not None:
@@ -907,7 +583,7 @@ class TrackExtractor:
         average_delta = float(np.mean(np.abs(delta)))
 
         # take half the max filtered value as a threshold
-        threshold = float(np.percentile(np.reshape(filtered, [-1]), q=TrackExtractor.THRESHOLD_PERCENTILE) / 2)
+        threshold = float(np.percentile(np.reshape(filtered, [-1]), q=self.config.threshold_percentile) / 2)
 
         # cap the threshold to something reasonable
         if threshold < self.min_threshold:
@@ -925,7 +601,6 @@ class TrackExtractor:
 
         return background, background_stats
 
-
 def apply_threshold(frame, threshold):
     """
     Creates a binary mask out of an image by applying a threshold.
@@ -936,3 +611,17 @@ def apply_threshold(frame, threshold):
     thresh[thresh < 0] = 0
     thresh[thresh > 0] = 1
     return thresh
+
+
+
+
+
+class BackgroundAnalysis:
+    """ Stores background analysis statistics. """
+    def __init__(self):
+        self.threshold = None
+        self.average_delta = None
+        self.max_temp = None
+        self.min_temp = None
+        self.mean_temp = None
+        self.background_deviation = None
